migrate individual functions and test for generateChartOptions.js
Accoriding to colocation principles, also all individual functions and tests for generateChartOptions.js were migrated to its own file as those are only used there. All individual external functions and test files were removed.
