--- src/functions/getChartData.js
@@ -1,60 +1,0 @@
-import Highcharts from "highcharts";
-
-export const getChartData = (impressions, clicks, title) =>
-  // Recives series of impressions, series of clicks and title message for Highcharts as input.
-  // Outputs options to visualize data on a HighchartsReact component.
-  ({
-    chart: {
-      type: "spline",
-      zoomType: "x"
-    },
-    title: {
-      text: title,
-      align: "left"
-    },
-    xAxis: {
-      type: "datetime",
-      dateTimeLabelFormats: {
-        day: "%e. %b"
-      }
-    },
-    yAxis: [
-      {
-        // Primary yAxis
-        labels: {
-          format: "{value}",
-          style: {
-            color: Highcharts.getOptions().colors[0]
-          }
-        },
-        title: {
-          text: "Clicks",
-          style: {
-            color: Highcharts.getOptions().colors[0]
-          }
-        },
-        opposite: false
-      },
-      {
-        // Secondary yAxis
-        gridLineWidth: 0,
-        title: {
-          text: "Impressions",
-          style: {
-            color: Highcharts.getOptions().colors[1]
-          }
-        },
-        labels: {
-          format: "{value}",
-          style: {
-            color: Highcharts.getOptions().colors[1]
-          }
-        },
-        opposite: true
-      }
-    ],
-    legend: {
-      enabled: true
-    },
-    series: [clicks, impressions]
-  });
