--- src/functions/tests/getTitleList.test.js
@@ -1,7 +1,0 @@
-import { getTitleList } from "../functions/getTitleList";
-
-test("getTitleList", () => {
-  expect(getTitleList(["Facebook Ads", "Google Adwords"])).toStrictEqual(
-    '"Facebook Ads" and "Google Adwords" '
-  );
-});
